fix(Web-UI): Fixed going back when viewing a tv show to the media overview not loading the last page correctly
