+++ src/WebAPI/ClientApp/src/components/MediaOverview/MediaOverview.vue
@@ -27,9 +27,9 @@
 		<!--	Overview bar	-->
 		<MediaOverviewBar
 			:server="
-				libraryStore.getServerByLibraryId(props.libraryId)"
+				libraryStore.getServerByLibraryId(libraryId)"
 			:library="library"
-			:detail-mode="!mediaOverviewStore.showMediaOverview"
+			:detail-mode="false"
 			@back="closeDetailsOverview"
 			@view-change="changeView"
 			@selection-dialog="useOpenControlDialog(mediaSelectionDialogName)"
@@ -47,7 +47,7 @@
 				<QRow
 					id="media-container"
 					align="start">
-					<QCol v-show="mediaOverviewStore.showMediaOverview">
+					<QCol>
 						<template v-if="mediaOverviewStore.getMediaViewMode === ViewMode.Table">
 							<MediaTable
 								:rows="mediaOverviewStore.getMediaItems"
@@ -64,7 +64,7 @@
 						</template>
 					</QCol>
 					<!-- Alphabet Navigation -->
-					<AlphabetNavigation v-show="mediaOverviewStore.showMediaOverview" />
+					<AlphabetNavigation />
 				</QRow>
 			</template>
 		</template>
@@ -87,8 +87,6 @@
 				</QCol>
 			</QRow>
 		</template>
-		<!-- Media Details Display -->
-		<DetailsOverview :name="mediaDetailsDialogName" />
 		<!-- Media Selection Dialog -->
 		<MediaSelectionDialog :name="mediaSelectionDialogName" />
 		<!--	Loading overlay	-->
@@ -104,15 +102,13 @@
 import Log from 'consola';
 import { get, set } from '@vueuse/core';
 import { useSubscription } from '@vueuse/rxjs';
-import { useRouter, type RouteLocationNormalized, type RouteLocationNormalizedLoaded } from 'vue-router';
 import { type DownloadMediaDTO, type LibraryProgress, PlexMediaType, ViewMode } from '@dto';
-import { listenMediaOverviewOpenDetailsCommand, sendMediaOverviewOpenDetailsCommand } from '@composables/event-bus';
+import { sendMediaOverviewOpenDetailsCommand } from '@composables/event-bus';
 import {
 	useMediaOverviewBarDownloadCommandBus,
 	useMediaOverviewSortBus,
 	useOpenControlDialog,
 	listenMediaOverviewDownloadCommand,
-	useCloseControlDialog,
 	sendMediaOverviewDownloadCommand,
 	useMediaStore,
 	useMediaOverviewStore,
@@ -130,12 +126,10 @@ const mediaOverviewStore = useMediaOverviewStore();
 const downloadStore = useDownloadStore();
 const libraryStore = useLibraryStore();
 const serverStore = useServerStore();
-const router = useRouter();
 
 // endregion
 
 const downloadConfirmationName = 'mediaDownloadConfirmation';
-const mediaDetailsDialogName = 'mediaDetailsDialogName';
 const mediaSelectionDialogName = 'mediaSelectionDialogName';
 const isRefreshing = ref(false);
 
@@ -145,7 +139,6 @@ const loading = ref(false);
 
 const props = defineProps<{
 	libraryId: number;
-	mediaId: number;
 	mediaType: PlexMediaType;
 }>();
 
@@ -251,122 +244,21 @@ listenMediaOverviewDownloadCommand((command) => {
 	}
 });
 
-listenMediaOverviewOpenDetailsCommand((mediaId: number) => {
-	if (!mediaId) {
-		Log.error('mediaId was invalid, could not open details', mediaId);
-		return;
-	}
-
-	// Replace the url with the library id and media id
-	router
-		.push({
-			name: 'details-overview',
-			params: { libraryId: props.libraryId, tvShowId: mediaId },
-		})
-		.then(() => {
-			useOpenControlDialog(mediaDetailsDialogName, { mediaId, type: props.mediaType });
-			mediaOverviewStore.showMediaOverview = false;
-		});
-});
-
-function closeDetailsOverview() {
-	Log.info('closeDetailsOverview');
-	// Replace the url with the library id
-	router
-		.push({
-			name: 'media-overview',
-			params: { libraryId: props.libraryId },
-		})
-		.then(() => {
-			useCloseControlDialog(mediaDetailsDialogName);
-			mediaOverviewStore.showMediaOverview = true;
-		});
-}
-
 useMediaOverviewBarDownloadCommandBus().on(() => {
-	if (mediaOverviewStore.showMediaOverview) {
-		const downloadCommand: DownloadMediaDTO = {
-			plexServerId: libraryStore.getServerByLibraryId(props.libraryId)?.id ?? 0,
-			plexLibraryId: props.libraryId,
-			mediaIds: mediaOverviewStore.selection.keys,
-			type: props.mediaType,
-		};
-		sendMediaOverviewDownloadCommand([downloadCommand]);
-	}
+	const downloadCommand: DownloadMediaDTO = {
+		plexServerId: libraryStore.getServerByLibraryId(props.libraryId)?.id ?? 0,
+		plexLibraryId: props.libraryId,
+		mediaIds: mediaOverviewStore.selection.keys,
+		type: props.mediaType,
+	};
+	sendMediaOverviewDownloadCommand([downloadCommand]);
 });
 
 useMediaOverviewSortBus().on((event) => {
 	mediaOverviewStore.sortMedia(event);
 });
 
-function setupRouter() {
-	router.beforeEach((to, from, next) => {
-		// From MediaOverview => DetailsOverview
-		if (!from.path.includes('details') && to.path.includes('details')) {
-			let tableRef: HTMLElement | null = null;
-			if (mediaOverviewStore.getMediaViewMode === ViewMode.Table) {
-				tableRef = document.getElementById('media-table-scroll');
-			}
-			if (mediaOverviewStore.getMediaViewMode === ViewMode.Poster) {
-				tableRef = document.getElementById('poster-table');
-			}
-
-			if (!tableRef) {
-				Log.error('tableRef was null for type', mediaOverviewStore.getMediaViewMode);
-				return next();
-			}
-			// Save the current scroll position to be restored when navigating back
-			to.meta?.scrollPos && (to.meta.scrollPos.top = tableRef.scrollTop);
-		}
-
-		return next();
-	});
-
-	router.options.scrollBehavior = (to: RouteLocationNormalized, from: RouteLocationNormalizedLoaded) => {
-		// From DetailsOverview => MediaOverview
-		if (from.path.includes('details') && !to.path.includes('details')) {
-			return new Promise((resolve) => {
-				setTimeout(() => {
-					let tableRef: HTMLElement | null = null;
-					switch (mediaOverviewStore.getMediaViewMode) {
-						case ViewMode.Table:
-							tableRef = document.getElementById('media-table-scroll');
-							break;
-						case ViewMode.Poster:
-							tableRef = document.getElementById('poster-table');
-							break;
-						default:
-							Log.error('Unknown mediaViewMode', mediaOverviewStore.getMediaViewMode);
-							return;
-					}
-
-					if (!tableRef) {
-						Log.error('tableRef was null for type', mediaOverviewStore.getMediaViewMode);
-						return;
-					}
-
-					tableRef.scrollTo({
-						behavior: 'smooth',
-						top: from.meta.scrollPos?.top ?? -1,
-						left: 0,
-					});
-
-					return resolve({
-						behavior: 'smooth',
-						top: from.meta.scrollPos?.top ?? -1,
-						left: 0,
-					});
-				}, 100);
-			});
-		}
-		return Promise.resolve();
-	};
-}
-
-// endregion
-
 onMounted(() => {
-	Log.info('MediaOverview => onMounted');
 	resetProgress(false);
 	set(isRefreshing, false);
 
@@ -380,8 +272,6 @@ onMounted(() => {
 		page: 0,
 		size: 0,
 	});
-
-	setupRouter();
 
 	useSubscription(
 		useSignalrStore()
