+++ src/WebAPI/ClientApp/src/pages/tvshows/[libraryId]/index.vue
